refactor: CustomAdapter と diffUtil について、OneFragment にまとまっていると読みづらいので別ファイル化した。
